fix(core): match Swift's %03d pad and 64-bit int width in the MA2 port (#827)
Turns the three cases added by the previous commit green.

`Ma2CueNumber.CommandString` spelled the sub number with `ToString("D3")`,
which pads the digits to three and prepends the sign; Swift's `%03d` pads
to a total width of three with the sign inside it. Spelled out as an
explicit `PadLeft(3 - sign.Length, '0')` so the width arithmetic is
visible rather than delegated to a format string that means something
else.

The three `(int)` casts over `Math.Round` now cast to `long`. Swift's
`Int` is 64-bit and .NET Core 3.0+ saturates a narrowing float→int cast
at 2147483647 instead of trapping, so a large cue number or cue time
would have clamped silently on Windows while macOS kept counting. Only
the cue-number site is vector-pinned; the `Ma2TrigTime` and
`Ma2TimecodeXmlGenerator` sites need a cue time near 1e8 seconds to
diverge, so they carry the threshold in their remarks instead.

Also routes the two `At Exec {page}.{number}` interpolations through
`CultureInfo.InvariantCulture`, matching every other integer in both
planners. No behavioural change — `Ma2PushTarget.IsValid` rejects the
negative values a non-ASCII `NegativeSign` culture could alter.

+++ windows/OnlyCue.Core.Tests/Ma2Vector.cs
@@ -68,8 +68,8 @@ public sealed record Ma2TrigTimeCase(
 public sealed record Ma2CueNumberCase(
     [property: JsonPropertyName("name")] string Name,
     [property: JsonPropertyName("value")] string Value,
-    [property: JsonPropertyName("expectNumber")] int ExpectNumber,
-    [property: JsonPropertyName("expectSubNumber")] int ExpectSubNumber,
+    [property: JsonPropertyName("expectNumber")] long ExpectNumber,
+    [property: JsonPropertyName("expectSubNumber")] long ExpectSubNumber,
     [property: JsonPropertyName("expectCommandString")] string ExpectCommandString);
 
 public sealed record Ma2NameCase(
